+++ example/lib/components/app/app_component.dart
@@ -1,8 +1,9 @@
 /*
  * fluro
- * A Posse Production
- * http://goposse.com
- * Copyright (c) 2018 Posse Productions LLC. All rights reserved.
+ * Created by Yakka
+ * https://theyakka.com
+ * 
+ * Copyright (c) 2018 Yakka, LLC. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
 import '../../config/application.dart';
@@ -28,6 +29,7 @@ class AppComponentState extends State<AppComponent> {
   Widget build(BuildContext context) {
     final app = new MaterialApp(
       title: 'Fluro',
+      debugShowCheckedModeBanner: false,
       theme: new ThemeData(
         primarySwatch: Colors.blue,
       ),